Shared IOApp test again
- Register it to make ScalaJS happy
- Back to Fibonacci as Thread.sleep doesn't work on scalaJS
- Don't use Runtime for cross-building reasons
- Fewer fibers to try to strike a balance where it should fail on JVMs
  that don't have a ludicrous number of cpus available but also
  completes within my lifetime on JS
